WIP -- Modularize constrained optimizers
