--- cooper/constrained_optimizer/__init__.py
@@ -1,1 +1,0 @@
-from .constrained_optimizer import ConstrainedOptimizer
