+++ cooper/optim/__init__.py
@@ -1,2 +1,7 @@
+from .constrained_optimizers import *
+from .constrained_optimizers.utils import (
+    create_optimizer_from_kwargs,
+    load_cooper_optimizer_from_state_dict,
+)
 from .extra_optimizers import ExtraAdam, ExtragradientOptimizer, ExtraSGD
 from .utils import partial_optimizer, partial_scheduler
