+++ docs/source/optim.rst
@@ -8,7 +8,7 @@ Optim Module
 Partial optimizer instantiation
 -------------------------------
 
-When constructing a :py:class:`~cooper.constrained_optimizer.ConstrainedOptimizer`, the
+When constructing a :py:class:`~cooper.optim.constrained_optimizer.ConstrainedOptimizer`, the
 ``dual_optimizer`` parameter is expected to be a
 :py:class:`torch.optim.Optimizer` for which the ``params`` argument has **not
 yet** been passed. The rest of the instantiation of the ``dual_optimizer`` is
@@ -66,7 +66,7 @@ Dual learning rate scheduler
 .. _dual_lr_scheduler:
 
 When constructing a
-:py:class:`~cooper.constrained_optimizer.ConstrainedOptimizer`,
+:py:class:`~cooper.optim.constrained_optimizer.ConstrainedOptimizer`,
 the ``dual_scheduler`` parameter is expected to be a *partially instantiated*
 learning rate scheduler from Pytorch, for which the ``optimizer`` argument has
 **not yet** been passed. The :py:meth:`cooper.optim.partial_scheduler` method
@@ -79,7 +79,7 @@ managed internally by **Cooper**.
     The call to the ``step()`` method of the dual optimizer is handled
     internally by **Cooper**. However, you must perform the call to the dual
     scheduler's ``step`` method manually. This will usually come after several
-    calls to :py:meth:`cooper.constrained_optimizer.ConstrainedOptimizer.step`.
+    calls to :py:meth:`cooper.optim.constrained_optimizer.ConstrainedOptimizer.step`.
 
     The reasoning behind this design is to provide you, the user, with greater
     visibility and control on the dual learning rate scheduler. For example, you
@@ -140,7 +140,7 @@ extra-gradient in the context of solving Variational Inequality Problems.
 .. warning::
 
     If you decide to use extra-gradient optimizers for defining a
-    :py:class:`~cooper.constrained_optimizer.ConstrainedOptimizer`, the primal
+    :py:class:`~cooper.optim.constrained_optimizer.ConstrainedOptimizer`, the primal
     and dual optimizers must **both** be instances of classes inheriting from
     :py:class:`ExtragradientOptimizer`.
 
@@ -149,7 +149,7 @@ extra-gradient in the context of solving Variational Inequality Problems.
 
     Due to the calculation of gradients at the "look-ahead" point
     :math:`\omega_{t+1/2}`, the call to
-    :py:meth:`cooper.constrained_optimizer.ConstrainedOptimizer.step` requires
+    :py:meth:`cooper.optim.constrained_optimizer.ConstrainedOptimizer.step` requires
     passing the parameters needed for the computation of the
     :py:meth:`cooper.problem.ConstrainedMinimizationProblem.closure`.
 
@@ -162,7 +162,7 @@ extra-gradient in the context of solving Variational Inequality Problems.
 
         model = ...
 
-        cmp = cooper.ConstrainedMinimizationProblem(is_constrained=True)
+        cmp = cooper.ConstrainedMinimizationProblem()
         formulation = cooper.Formulation(...)
 
         # Non-extra-gradient optimizers
