+++ docs/source/problem.rst
@@ -94,7 +94,7 @@ define a ``ConstrainedMinimizationProblem`` in **Cooper**.
         def __init__(self, problem_attributes, criterion):
             self.problem_attributes = problem_attributes
             self.criterion = criterion
-            super().__init__(is_constrained=True)
+            super().__init__()
 
         def closure(self, model, inputs, targets):
