+++ tests/test_lagrangian_formulation.py
@@ -10,7 +10,7 @@ import cooper
 def test_lagrangian_formulation():
     class DummyCMP(cooper.ConstrainedMinimizationProblem):
         def __init__(self):
-            super().__init__(is_constrained=True)
+            super().__init__()
 
         def closure(self):
             pass
